+++ Workade/Views&ViewModels/NearbyPlace/NearbyPlaceDetailView.swift
@@ -11,9 +11,11 @@ class NearbyPlaceDetailView: UIView {
     let officeModel: OfficeModel
     var introduceViewModel: IntroduceViewModel
     
+    var mapVC: MapViewController!
     var introduceBottomConstraints: NSLayoutConstraint!
     var galleryBottomConstraints: NSLayoutConstraint!
     var featureBottomConstraints: NSLayoutConstraint!
+    var mapBottomConstrains: NSLayoutConstraint!
     
     let contensContainerView: UIView = {
         let contensContainerView = UIView()
@@ -45,23 +47,38 @@ class NearbyPlaceDetailView: UIView {
         return view
     }()
     
+    var mapView: UIView!
+    
     init(officeModel: OfficeModel) {
         self.officeModel = officeModel
         self.introduceViewModel = IntroduceViewModel()
+        
         super.init(frame: .zero)
-        
-        introduceBottomConstraints = introduceView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor, constant: -20)
-        galleryBottomConstraints = galleryView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor, constant: -20)
-        featureBottomConstraints = featureView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor, constant: -20)
-        
         translatesAutoresizingMaskIntoConstraints = false
-        
+        setupMapView()
+        setupViewBottomConstraints()
         setupIntroduceView()
         setupLayout()
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+    
+    func setupMapView() {
+        mapVC = MapViewController(office: officeModel)
+        mapVC.setMapCamera()
+        
+        mapView = mapVC.view
+        mapView.isHidden = true
+        mapView.translatesAutoresizingMaskIntoConstraints = false
+    }
+    
+    func setupViewBottomConstraints() {
+        introduceBottomConstraints = introduceView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor, constant: -20)
+        galleryBottomConstraints = galleryView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor, constant: -20)
+        featureBottomConstraints = featureView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor, constant: -20)
+        mapBottomConstrains = mapView.bottomAnchor.constraint(equalTo: contensContainerView.bottomAnchor)
     }
     
     func setupLayout() {
@@ -87,7 +104,7 @@ class NearbyPlaceDetailView: UIView {
             featureView.topAnchor.constraint(equalTo: contensContainerView.topAnchor, constant: 30),
             featureView.leadingAnchor.constraint(equalTo: contensContainerView.leadingAnchor, constant: 20),
             featureView.trailingAnchor.constraint(equalTo: contensContainerView.trailingAnchor, constant: -20),
-            featureView.heightAnchor.constraint(equalToConstant: UIScreen.main.bounds.height - .topSafeArea - 50)
+            featureView.heightAnchor.constraint(equalToConstant: UIScreen.main.bounds.height - .topSafeArea - 52)
         ])
         
         contensContainerView.addSubview(galleryView)
@@ -95,7 +112,16 @@ class NearbyPlaceDetailView: UIView {
             galleryView.topAnchor.constraint(equalTo: contensContainerView.topAnchor, constant: 0),
             galleryView.leadingAnchor.constraint(equalTo: contensContainerView.leadingAnchor, constant: 0),
             galleryView.trailingAnchor.constraint(equalTo: contensContainerView.trailingAnchor, constant: 0),
-            galleryView.heightAnchor.constraint(equalToConstant: UIScreen.main.bounds.height - .topSafeArea - 50)
+            galleryView.heightAnchor.constraint(equalToConstant: UIScreen.main.bounds.height - .topSafeArea - 52)
+        ])
+        
+        contensContainerView.addSubview(mapView)
+        // 네비게이션 바 높이 = 60으로 고정값임. 디바이스별로 구분할 필요 없음.
+        NSLayoutConstraint.activate([
+            mapView.topAnchor.constraint(equalTo: contensContainerView.topAnchor),
+            mapView.leadingAnchor.constraint(equalTo: contensContainerView.leadingAnchor),
+            mapView.trailingAnchor.constraint(equalTo: contensContainerView.trailingAnchor),
+            mapView.heightAnchor.constraint(equalToConstant: UIScreen.main.bounds.height - .topSafeArea - 52 - 60)
         ])
     }
 }
